♻️ [refactor]: 공통으로 쓰이는 enum 패키지 변경
