+++ src/main/java/com/PickOne/domain/application/dto/request/ApplicationInstrumentDto.java
@@ -1,7 +1,7 @@
 package com.PickOne.domain.application.dto.request;
 
-import com.PickOne.domain.recruitments.model.Instrument;
-import com.PickOne.domain.recruitments.model.Proficiency;
+import com.PickOne.global.common.enums.Instrument;
+import com.PickOne.global.common.enums.Proficiency;
 import com.PickOne.domain.recruitments.model.entity.Recruitment;
 import com.PickOne.domain.recruitments.model.entity.RecruitmentInstrument;
